requirements.txt
aiohttp==3.7.4.post0
astroid==2.7.3
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.6
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cupy-cuda111==9.4.0
cymem==2.0.5
decorator==5.0.9
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.1.0/en_core_web_md-3.1.0-py3-none-any.whl
fastrlock==0.6
google-api-core==2.0.1
google-auth==2.0.2
google-auth-oauthlib==0.4.6
google-cloud-aiplatform==1.4.1
google-cloud-bigquery==2.26.0
google-cloud-core==2.0.0
google-cloud-storage==1.42.1
google-crc32c==1.1.2
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
grpcio==1.40.0
idna==3.2
isort==5.9.3
Jinja2==3.0.1
joblib==1.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==5.1.0
murmurhash==1.0.5
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.2
pathy==0.6.0
Pillow==8.3.2
platformdirs==2.3.0
preshed==3.0.5
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pylint==2.10.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
smart-open==5.2.1
spacy==3.1.2
spacy-legacy==3.0.8
srsly==2.4.1
thinc==8.0.10
threadpoolctl==2.2.0
torch==1.8.0+cu111
torchtext=
tokenizers==0.10.0
toml==0.10.2
tqdm==4.62.2
typed-ast==1.4.3
typer==0.3.2
typing-extensions==3.10.0.2
urllib3==1.26.6
wasabi==0.8.2
wrapt==1.12.1
yarl==1.6.3
zipp==3.5.0
